+++ src/tokens/ios/MuiBrandTokens.swift
@@ -2,7 +2,7 @@
 // MuiBrandTokens.swift
 //
 
-// Don't edit directly • Generated on Tue, 02 Jun 2026 14:17:31 GMT • muibook.com
+// Don't edit directly • Generated on Tue, 02 Jun 2026 15:06:48 GMT • muibook.com
 
 import UIKit